fix: correct algorithm errors in 8 kernel implementations
- volk_16ic_magnitude_16i: precompute invScalarF, multiply instead of divide
- volk_32f_x2_pow_32f: rename _sse4_1 to _sse2 (no SSE4.1 intrinsics used)
- volk_32f_x3_sum_of_poly_32f: rename _a_avx2_fma to _a_avx_fma (guard is AVX+FMA)
- volk_32fc_deinterleave_64f_x2: rename _neon to _neonv8 (uses ARMv8-A features)
- volk_32fc_deinterleave_real_64f: rename _neon to _neonv8 (uses ARMv8-A features)
- volk_32fc_s32f_x2_power_spectral_density_32f: 1.0 → 1.0f, sqrt → sqrtf
- volk_32fc_s32fc_x2_rotator_32fc: fix include guards to include _x2_
- volk_64u_popcnt: fix #endif comment to match full guard condition

+++ kernels/volk/volk_32fc_s32f_x2_power_spectral_density_32f.h
@@ -88,9 +88,9 @@ volk_32fc_s32f_x2_power_spectral_density_32f_generic(float* logPowerOutput,
                                                      const float rbw,
                                                      unsigned int num_points)
 {
-    if (rbw != 1.0)
+    if (rbw != 1.0f)
         volk_32fc_s32f_power_spectrum_32f(
-            logPowerOutput, complexFFTInput, normalizationFactor * sqrt(rbw), num_points);
+            logPowerOutput, complexFFTInput, normalizationFactor * sqrtf(rbw), num_points);
     else
         volk_32fc_s32f_power_spectrum_32f(
             logPowerOutput, complexFFTInput, normalizationFactor, num_points);
